refactor: split export/pandas_engine.py into {format}_{engine}.py files, add N-Quads export
- Split monolithic pandas_engine.py into:
  - excel_pandas.py (openpyxl-based Excel export)
  - csv_pandas.py (pandas CSV export)
  - cimxml_pandas.py (lxml-based CIM XML generation)
  - networkx_pandas.py (NetworkX graph export)
- Add N-Quads export:
  - nquads_polars.py (vectorized, 3.2M rows/s)
  - nquads_pandas.py (pandas fallback, no polars dep)
- Add cimxml_cython_pugixml.pyx (Arrow -> pugixml C++ export from dev_shacl)
- Dispatcher auto-selects polars for nquads when available
- Register export_to_nquads on pandas DataFrame and accessor
